+++ src/main.cpp
@@ -70,7 +70,7 @@ int main(int argc, char* argv[]){
         CPU_time = (double)(clock() - START) / CLOCKS_PER_SEC;
         REAL_time = finish.tv_sec - start.tv_sec + (finish.tv_nsec - start.tv_nsec) / 1000000000.0;
 
-        int CPU_int, CPU_dec, REAL_int, REAL_dec;
+        size_t CPU_int, CPU_dec, REAL_int, REAL_dec;
 
         CPU_int = floor(log10(CPU_time)) + 1;
         CPU_dec = 3 - CPU_int;
